backend test/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.0
python-dotenv==1.0.0
pandas==2.1.3
numpy==1.24.3
scapy==2.5.0
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.0
Pillow==10.1.0
celery==5.3.4
redis==5.0.1
psycopg2-binary==2.9.9 